Release-1.6.4 Testing
This patch corrects a number of issues found in developer testing
prior to the release-1.6.4 code freeze. Thus, a number of tests
have been changed to work correctly with server 5.7.12.

Additional changes to the library code include the following

- allow ignore of fatal error in kill() when process has
  already stopped for serverclone.py. This solves the issue
  of 5.7+ servers delay between mysqladmin shutdown command
  and actual shutdown of the server
- add socket for 5.7+ servers in serverinfo.py restart from
  downed server state (required for 5.7+)
- ignore timeout option for utilities that don't use it in
  topology.py
- added a new method to utilities.py for killing processes
  (a variant of the code in mutlib)

+++ mysql/utilities/command/serverclone.py
@@ -1,5 +1,5 @@
 #
-# Copyright (c) 2010, 2015, Oracle and/or its affiliates. All rights reserved.
+# Copyright (c) 2010, 2016, Oracle and/or its affiliates. All rights reserved.
 #
 # This program is free software; you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -307,9 +307,10 @@ def clone_server(conn_val, options):
         if os.name == "posix":
             try:
                 os.kill(proc.pid, subprocess.signal.SIGTERM)
-            except OSError:
-                raise UtilError("Failed to kill process with pid '{0}'"
-                                "".format(proc.pid))
+            except OSError as error:
+                if not error.strerror.startswith("No such process"):
+                    raise UtilError("Failed to kill process with pid '{0}'"
+                                    "".format(proc.pid))
         else:
             ret_code = subprocess.call("taskkill /F /T /PID "
                                        "{0}".format(proc.pid), shell=True)
